server/requirements.txt
fastapi
uvicorn
pydantic
openai
sentence-transformers
torch
faiss-cpu
lxml
numpy
tqdm
python-dotenv
